Refactor geo module
