+++ src/geo/error.rs
@@ -1,0 +1,10 @@
+use maxminddb::MaxMindDBError;
+use thiserror::Error;
+
+#[derive(Error, Debug)]
+pub enum GeoError {
+    #[error("continent is not recognised")]
+    ContinentUnknown,
+    #[error(transparent)]
+    MaxMindDBError(#[from] MaxMindDBError),
+}
